Crate snowbridge_pallet_system

Governance API for controlling the Ethereum side of the bridge

Extrinsics

Agents

Agents are smart contracts on Ethereum that act as proxies for consensus systems on Polkadot networks.

Call::create_agent: Create agent for a sibling parachain

Call::transfer_native_from_agent: Withdraw ether from an agent

The create_agent extrinsic should be called via an XCM Transact instruction from the sibling parachain.

Channels

Each sibling parachain has its own dedicated messaging channel for sending and receiving messages. As a prerequisite to creating a channel, the sibling should have already created an agent using the create_agent extrinsic.

Call::create_channel: Create channel for a sibling

Call::update_channel: Update a channel for a sibling

Governance

Only Polkadot governance itself can call these extrinsics. Delivery fees are waived.

Call::upgrade`: Upgrade the gateway contract

Call::set_operating_mode: Update the operating mode of the gateway contract

Call::force_update_channel: Allow root to update a channel for a sibling

Call::force_transfer_native_from_agent: Allow root to withdraw ether from an agent

Typically, Polkadot governance will use the force_transfer_native_from_agent and force_update_channel and extrinsics to manage agents and channels for system parachains.

Re-exports

pub use pallet::*;

pub use weights::*;

pub use pallet::*;

Modules

api

Helpers for implementing runtime api

migration

Governance API for controlling the Ethereum side of the bridge

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for snowbridge_system

Enums

PaysFee

Whether a fee should be withdrawn to an account for sending an outbound message

Type Aliases

AccountIdOf

BalanceOf

PricingParametersOf